+++ skforecast/ForecasterAutoregDirect/ForecasterAutoregDirect.py
@@ -1447,7 +1447,8 @@ class ForecasterAutoregDirect(ForecasterBase):
  
     def get_feature_importances(
         self, 
-        step: int
+        step: int,
+        sort_importance: bool=True
     ) -> pd.DataFrame:
         """
         Return feature importance of the model stored in the forecaster for a
@@ -1462,6 +1463,8 @@ class ForecasterAutoregDirect(ForecasterBase):
         step : int
             Model from which retrieve information (a separate model is created 
             for each forecast time step). First step is 1.
+        sort_importance: bool, default `True`
+            If `True`, sorts the feature importances in descending order.
 
         Returns
         -------
@@ -1472,7 +1475,7 @@ class ForecasterAutoregDirect(ForecasterBase):
 
         if not isinstance(step, int):
             raise TypeError(
-                f'`step` must be an integer. Got {type(step)}.'
+                f"`step` must be an integer. Got {type(step)}."
             )
         
         if not self.fitted:
@@ -1495,7 +1498,8 @@ class ForecasterAutoregDirect(ForecasterBase):
         idx_columns_lags = np.arange(len(self.lags))
         if self.included_exog:
             idx_columns_exog = np.flatnonzero(
-                                [name.endswith(f"step_{step}") for name in self.X_train_col_names]
+                                   [name.endswith(f"step_{step}") 
+                                    for name in self.X_train_col_names]
                                )
         else:
             idx_columns_exog = np.array([], dtype=int)
@@ -1522,5 +1526,9 @@ class ForecasterAutoregDirect(ForecasterBase):
                                       'feature': feature_names,
                                       'importance': feature_importances
                                   })
+            if sort_importance:
+                feature_importances = feature_importances.sort_values(
+                                          by='importance', ascending=False
+                                      )
 
         return feature_importances