+++ skforecast/ForecasterAutoregMultiSeriesCustom/tests/test_get_feature_importances.py
@@ -59,7 +59,7 @@ def test_output_get_feature_importances_when_regressor_is_RandomForest():
                  )
     forecaster.fit(series=series)
 
-    results = forecaster.get_feature_importances()
+    results = forecaster.get_feature_importances(sort_importance=False)
     expected = pd.DataFrame({
                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
                    'importance': np.array([0.9446366782006932, 0.0, 0.05536332179930687, 0.0, 0.0])
@@ -85,7 +85,7 @@ def test_output_get_feature_importances_when_regressor_is_RandomForest_with_exog
     expected = pd.DataFrame({
                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', 'exog', '1', '2'],
                    'importance': np.array([0.73269896, 0., 0.21193772, 0.05536332, 0., 0.])
-               })
+               }).sort_values(by='importance', ascending=False)
 
     pd.testing.assert_frame_equal(results, expected)
 
@@ -102,7 +102,7 @@ def test_output_get_feature_importances_when_regressor_is_LinearRegression():
                  )
     forecaster.fit(series=series)
 
-    results = forecaster.get_feature_importances()
+    results = forecaster.get_feature_importances(sort_importance=False)
     expected = pd.DataFrame({
                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
                    'importance': np.array([3.33333333e-01, 3.33333333e-01, 3.33333333e-01, 
@@ -127,7 +127,7 @@ def test_output_get_feature_importances_when_regressor_is_LinearRegression_with_
                  )
     forecaster.fit(series=series_2, exog=pd.Series(np.arange(10, 15), name='exog'))
 
-    results = forecaster.get_feature_importances()
+    results = forecaster.get_feature_importances(sort_importance=False)
     expected = pd.DataFrame({
                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', 'exog', '1', '2'],
                    'importance': np.array([2.00000000e-01,  2.00000000e-01, 2.00000000e-01,  
@@ -180,7 +180,7 @@ def test_output_get_feature_importances_when_pipeline_LinearRegression():
                  )
     forecaster.fit(series=series)
 
-    results = forecaster.get_feature_importances()
+    results = forecaster.get_feature_importances(sort_importance=False)
     expected = pd.DataFrame({
                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
                    'importance': np.array([6.66666667e-01, 6.66666667e-01, 6.66666667e-01, 
@@ -205,7 +205,7 @@ def test_output_get_feature_importances_when_pipeline_RandomForestRegressor():
                  )
     forecaster.fit(series=series)
 
-    results = forecaster.get_feature_importances()
+    results = forecaster.get_feature_importances(sort_importance=False)
     expected = pd.DataFrame({
                    'feature': ['custom_predictor_0', 'custom_predictor_1', 'custom_predictor_2', '1', '2'],
                    'importance': np.array([0.9446366782006932, 0.0, 0.05536332179930687, 0.0, 0.0])