+++ skforecast/ForecasterAutoregMultiVariate/ForecasterAutoregMultiVariate.py
@@ -1623,7 +1623,8 @@ class ForecasterAutoregMultiVariate(ForecasterBase):
     
     def get_feature_importances(
         self,
-        step: int
+        step: int,
+        sort_importance: bool=True
     ) -> pd.DataFrame:
         """
         Return feature importance of the model stored in the forecaster for a
@@ -1638,6 +1639,8 @@ class ForecasterAutoregMultiVariate(ForecasterBase):
         step : int
             Model from which retrieve information (a separate model is created 
             for each forecast time step). First step is 1.
+        sort_importance: bool, default `True`
+            If `True`, sorts the feature importances in descending order.
 
         Returns
         -------
@@ -1648,7 +1651,7 @@ class ForecasterAutoregMultiVariate(ForecasterBase):
 
         if not isinstance(step, int):
             raise TypeError(
-                f'`step` must be an integer. Got {type(step)}.'
+                f"`step` must be an integer. Got {type(step)}."
             )
 
         if not self.fitted:
@@ -1672,8 +1675,8 @@ class ForecasterAutoregMultiVariate(ForecasterBase):
         idx_columns_lags = np.arange(len_columns_lags)
         if self.included_exog:
             idx_columns_exog = np.flatnonzero(
-                                [name.endswith(f"step_{step}")
-                                 for name in self.X_train_col_names]
+                                   [name.endswith(f"step_{step}")
+                                    for name in self.X_train_col_names]
                                )
         else:
             idx_columns_exog = np.array([], dtype=int)
@@ -1700,5 +1703,9 @@ class ForecasterAutoregMultiVariate(ForecasterBase):
                                       'feature': feature_names,
                                       'importance': feature_importances
                                   })
+            if sort_importance:
+                feature_importances = feature_importances.sort_values(
+                                          by='importance', ascending=False
+                                      )
 
         return feature_importances